+++ test/common.config.spec.js
@@ -1,6 +1,7 @@
 var CONF = require("../src_new/conf.js");
 var CONSTANTS = require("../src_new/constants.js");
 var COMMON_CONFIG = require("../src_new/common.config.js");
+var expect = require("chai").expect;
 
 describe('COMMON CONFIG FILE', function () {  
     describe('#getGdprActionTimeout', function () {
@@ -21,4 +22,56 @@ describe('COMMON CONFIG FILE', function () {
             done();
         });
     });
+
+    describe('setConsentConfig function', function () {
+        var prebidConfig = {};
+        var cmpApi = 'iab';
+        var timeout = 2000;
+
+        it('setConsentConfig a function', function (done) {
+            COMMON_CONFIG.setConsentConfig.should.be.a('function');
+            done();
+        });
+
+        it('should set consent management config  when its not present', function (done) {
+            var expPrebidConfig = {
+                consentManagement: {
+                    gpp: {
+                        cmpApi,
+                        timeout
+                    }
+                }
+            };
+            const actPrebifConfig = COMMON_CONFIG.setConsentConfig(prebidConfig, 'gpp', cmpApi, timeout);
+            expect(actPrebifConfig).to.be.deep.equal(expPrebidConfig);
+            done();
+        });
+
+        it('should set consent management config  when its present', function (done) {
+            var prebidConfig = {
+                consentManagement: {
+                    gdpr: {
+                        cmpApi,
+                        timeout
+                    }
+                }
+            }
+            var expPrebidConfig = {
+                consentManagement: {
+                    gdpr: {
+                        cmpApi,
+                        timeout
+                    },
+                    gpp: {
+                        cmpApi,
+                        timeout
+                    }
+                }
+            };
+            const actPrebifConfig = COMMON_CONFIG.setConsentConfig(prebidConfig, 'gpp', cmpApi, timeout);
+            expect(actPrebifConfig).to.be.deep.equal(expPrebidConfig);
+            done();
+        });
+    });
 });
+
